Minor reformatting.

+++ src/wrappers.c
@@ -16,10 +16,10 @@ function wrap_add(Left, Right : Ptr_64) return Unsigned_64;
 
 /* Used for both signed and unsigned */
 PLARGE_INTEGER wrap_add
-    (PLARGE_INTEGER result, PLARGE_INTEGER left, PLARGE_INTEGER right)
+(PLARGE_INTEGER result, PLARGE_INTEGER left, PLARGE_INTEGER right)
 {
-    *result = LargeIntegerAdd(*left, *right);
-    return result;
+  *result = LargeIntegerAdd (*left, *right);
+  return result;
 }
 
 /***************
@@ -30,10 +30,10 @@ function wrap_sub(Left, Right : Ptr_64) return Unsigned_64;
 
 /* Used for both signed and unsigned */
 PLARGE_INTEGER wrap_sub
-    (PLARGE_INTEGER result, PLARGE_INTEGER left, PLARGE_INTEGER right)
+(PLARGE_INTEGER result, PLARGE_INTEGER left, PLARGE_INTEGER right)
 {
-    *result = LargeIntegerSubtract(*left, *right);
-    return result;
+  *result = LargeIntegerSubtract(*left, *right);
+  return result;
 }
 
 /***************
@@ -41,10 +41,10 @@ function wrap_imul(Left, Right : Integer_32) return Integer_64;
 ****************/
 
 PLARGE_INTEGER wrap_imul
-    (PLARGE_INTEGER result, LONG left, LONG right)
+(PLARGE_INTEGER result, LONG left, LONG right)
 {
-    *result = EnlargedIntegerMultiply(left, right);
-    return result;
+  *result = EnlargedIntegerMultiply(left, right);
+  return result;
 }
 
 
@@ -53,10 +53,10 @@ function wrap_umul(Left, Right : Unsigned_32) return Unsigned_64;
 ****************/
 
 PLARGE_INTEGER wrap_umul
-    (PLARGE_INTEGER result, ULONG left, ULONG right)
+(PLARGE_INTEGER result, ULONG left, ULONG right)
 {
-    *result = EnlargedUnsignedMultiply(left, right);
-    return result;
+  *result = EnlargedUnsignedMultiply(left, right);
+  return result;
 }
 
 /***************
@@ -66,10 +66,10 @@ function wrap_imul64x32
 ****************/
 
 PLARGE_INTEGER wrap_imul64x32
-    (PLARGE_INTEGER result, PLARGE_INTEGER left, LONG right)
+(PLARGE_INTEGER result, PLARGE_INTEGER left, LONG right)
 {
-    *result = ExtendedIntegerMultiply(*left, right);
-    return result;
+  *result = ExtendedIntegerMultiply(*left, right);
+  return result;
 }
 
 /***************
@@ -77,10 +77,10 @@ function wrap_neg(Left : Ptr_64) return Integer_64;
 ****************/
 
 PLARGE_INTEGER wrap_neg
-    (PLARGE_INTEGER result, PLARGE_INTEGER left)
+(PLARGE_INTEGER result, PLARGE_INTEGER left)
 {
-    *result = LargeIntegerNegate(*left);
-    return result;
+  *result = LargeIntegerNegate(*left);
+  return result;
 }
 
 /***************
@@ -91,13 +91,13 @@ function wrap_ExtendedMagicDivide
 ****************/
 
 PLARGE_INTEGER wrap_ExtendedMagicDivide
-    (PLARGE_INTEGER result,
-    PLARGE_INTEGER Dividend,
-    PLARGE_INTEGER MagicDivisor,
-    CCHAR ShiftCount)
+(PLARGE_INTEGER result,
+ PLARGE_INTEGER Dividend,
+ PLARGE_INTEGER MagicDivisor,
+ CCHAR ShiftCount)
 {
-    *result = ExtendedMagicDivide(*Dividend, *MagicDivisor, ShiftCount);
-    return result;
+  *result = ExtendedMagicDivide(*Dividend, *MagicDivisor, ShiftCount);
+  return result;
 }
 
 /***************
@@ -108,13 +108,13 @@ function wrap_ExtendedLargeIntegerDivide
 ****************/
 
 PLARGE_INTEGER wrap_ExtendedLargeIntegerDivide
-    (PLARGE_INTEGER result,
-    PLARGE_INTEGER Dividend,
-    ULONG Divisor,
-    PULONG Remainder)
+(PLARGE_INTEGER result,
+ PLARGE_INTEGER Dividend,
+ ULONG Divisor,
+ PULONG Remainder)
 {
-    *result = ExtendedLargeIntegerDivide(*Dividend, Divisor, Remainder);
-    return result;
+  *result = ExtendedLargeIntegerDivide(*Dividend, Divisor, Remainder);
+  return result;
 }
 
 /***************
@@ -126,13 +126,13 @@ function wrap_LargeIntegerDivide
 ****************/
 
 PLARGE_INTEGER wrap_LargeIntegerDivide
-    (PLARGE_INTEGER result,
-    PLARGE_INTEGER Dividend,
-    PLARGE_INTEGER Divisor,
-    PLARGE_INTEGER Remainder)
+(PLARGE_INTEGER result,
+ PLARGE_INTEGER Dividend,
+ PLARGE_INTEGER Divisor,
+ PLARGE_INTEGER Remainder)
 {
-    *result = LargeIntegerDivide(*Dividend, *Divisor, Remainder);
-    return result;
+  *result = LargeIntegerDivide(*Dividend, *Divisor, Remainder);
+  return result;
 }
 
 /***************
@@ -141,12 +141,12 @@ function wrap_Shift_Left  (Value: Ptr_64; Amount: Natural)
 ****************/
 
 PLARGE_INTEGER wrap_Shift_Left
-    (PLARGE_INTEGER result,
-    PLARGE_INTEGER LargeInteger,
-    CCHAR ShiftCount)
+(PLARGE_INTEGER result,
+ PLARGE_INTEGER LargeInteger,
+ CCHAR ShiftCount)
 {
-    *result = LargeIntegerShiftLeft(*LargeInteger, ShiftCount);
-    return result;
+  *result = LargeIntegerShiftLeft(*LargeInteger, (int)ShiftCount);
+  return result;
 }
 
 /***************
@@ -155,12 +155,12 @@ function wrap_Shift_Right  (Value: Ptr_64; Amount: Natural)
 ****************/
 
 PLARGE_INTEGER wrap_Shift_Right
-    (PLARGE_INTEGER result,
-    PLARGE_INTEGER LargeInteger,
-    CCHAR ShiftCount)
+(PLARGE_INTEGER result,
+ PLARGE_INTEGER LargeInteger,
+ CCHAR ShiftCount)
 {
-    *result = LargeIntegerShiftRight(*LargeInteger, ShiftCount);
-    return result;
+  *result = LargeIntegerShiftRight(*LargeInteger, ShiftCount);
+  return result;
 }
 
 /***************
@@ -169,12 +169,12 @@ function wrap_Shift_Right_Arithmetic  (Value: Ptr_64; Amount: Natural)
 ****************/
 
 PLARGE_INTEGER wrap_Shift_Right_Arithmetic
-    (PLARGE_INTEGER result,
-    PLARGE_INTEGER LargeInteger,
-    CCHAR ShiftCount)
+(PLARGE_INTEGER result,
+ PLARGE_INTEGER LargeInteger,
+ CCHAR ShiftCount)
 {
-    *result = LargeIntegerArithmeticShift(*LargeInteger, ShiftCount);
-    return result;
+  *result = LargeIntegerArithmeticShift(*LargeInteger, ShiftCount);
+  return result;
 }
 
 /***************
@@ -183,12 +183,12 @@ function wrap_sImage (Value: Ptr_64;  Hex: Boolean := Standard.False)
 ****************/
 
 char *wrap_sImage
-    (char *result,
-    PLARGE_INTEGER Value,
-    long Hex)
+(char *result,
+ PLARGE_INTEGER Value,
+ long Hex)
 {
-    *result = 0;        /* Cause caller to raise execption */
-    return result;
+  *result = 0;        /* Cause caller to raise execption */
+  return result;
 }
 
 /***************
@@ -198,10 +198,10 @@ function wrap_uImage (Value: Ptr_64;  Hex: Boolean := Standard.False)
 ****************/
 
 char *wrap_uImage
-    (char *result,
-    PLARGE_INTEGER Value,
-    long Hex)
+(char *result,
+ PLARGE_INTEGER Value,
+ long Hex)
 {
-    *result = 0;        /* Cause caller to raise execption */
-    return result;
+  *result = 0;        /* Cause caller to raise execption */
+  return result;
 }
